Move home page files into app dir
