+++ app/blog/page.tsx
@@ -1,8 +1,8 @@
 import { getAllPosts } from "../../src/shared/posts.server";
 import { previewData } from "next/headers";
-import { HomeContainer } from "../../src/blog/index/home-container";
-import { Bio } from "../../src/blog/index/bio";
-import { PostList } from "../../src/blog/index/post-list";
+import { HomeContainer } from "./home-container";
+import { Bio } from "./bio";
+import { PostList } from "./post-list";
 
 export default async function Page() {
   const preview = previewData();
